project_aegis_phase1/requirements.txt
fastapi
uvicorn
pydantic
requests
python-dotenv
sqlalchemy
aiosqlite
streamlit
